AI Update
Now it seems to work almost correctly. It do not let win out of its hand to much anymore.
The problem is now to counter the ennemie.

Next step would be to make a play as the other player and evaluate what HIS victory would look like at X moment